Remove timeout for `wait-for` command
The default timeout for `wait-for` is 15 seconds.  If service startup happens to take a long time then the subscription setup portion fails but it doesn't shutdown the container or anything.  

I think just setting the timeout to infinite is probably a good enough solution since it's likely that the emulator will startup eventually.  A better solution could be to kill the emulator if it doesn't startup in time, but in that case I think you'd want the timeout to be configurable just in case someone has a REALLY slow startup.